split the rest part and use pandas
